pyproject.toml
[project]
name = "word-xml-python"
version = "0.1.0"
description = ""
authors = [
    {name = "TickHeart",email = "tickhere@qq.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lxml (>=6.0.2,<7.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "fastapi (>=0.121.1,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pdf2docx (>=0.5.8,<0.6.0)",
    "pyperclip (>=1.11.0,<2.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]

[tool.poetry]
packages = [{include = "word_xml_python", from = "src"}]


[tool.poetry.group.dev.dependencies]
types-lxml = "^2025.8.25"
ruff = "^0.14.8"
pre-commit = "^4.5.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
